refactor: the generated inline css placement for printing on front page
